requirements.txt
streamlit==1.38.0
sentence-transformers==3.1.1
torch==2.4.0
transformers==4.44.0
nltk==3.9.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
pdfplumber==0.11.0
PyPDF2==3.0.1
python-docx==1.1.2
mysql-connector-python==9.0.0
protobuf==5.27.2
altair==5.3.0
tqdm==4.66.5
requests==2.32.3